shared/ce: refactor challenge-grid
