+++ shared/ce/src/components/challenge-grid/test/challenge-grid.e2e.ts
